Move ProgressState to types/group
