+++ static/app/components/stream/group.tsx
@@ -34,6 +34,7 @@ import type {
   GroupReprocessing,
   InboxDetails,
   PriorityLevel,
+  ProgressState,
 } from 'sentry/types/group';
 import type {NewQuery} from 'sentry/types/organization';
 import type {User} from 'sentry/types/user';
@@ -66,7 +67,6 @@ import {
 import {
   formatProgressState,
   getProgressIcon,
-  type ProgressState,
 } from 'sentry/views/issueList/utils/progress';
 
 export const DEFAULT_STREAM_GROUP_STATS_PERIOD = '24h';
